[SGL-547] - Correção da tabela ator para ser criada com indentity e sequence de acordo com o dialect do banco.

+++ requirement/requirement-commons/src/main/resources/db/ddl/sqlserver/create-table-actor.sql
@@ -4,7 +4,7 @@
 /* Table: TB_ATOR                                               */
 /*==============================================================*/
 CREATE TABLE DBSINGULAR.TB_ATOR (
-   CO_ATOR  BIGINT       IDENTITY,
+   CO_ATOR  INTEGER     IDENTITY,
    CO_USUARIO           VARCHAR(60)          NOT NULL,
   CONSTRAINT PK_ATOR PRIMARY KEY (CO_ATOR)
 );
